+++ CHANGELOG.md
@@ -1,9 +1,16 @@
 # go-connect
 
+<a name="v1.0.1"></a>
+## [v1.0.1] - 2020-03-12
+### Fixes
+- fix goredis connect on non clusetered redis
+
+
 <a name="v1.0.0"></a>
 ## v1.0.0 - 2020-03-11
 ### New Features
-- init go-connect with http and redis connector
+- init go-connect with http and redis connector ([#1](https://github.com/kumparan/go-connect/issues/1))
 
 
-[Unreleased]: https://github.com/kumparan/go-connect/compare/v1.0.0...HEAD
+[Unreleased]: https://github.com/kumparan/go-connect/compare/v1.0.1...HEAD
+[v1.0.1]: https://github.com/kumparan/go-connect/compare/v1.0.0...v1.0.1